Fix mior issue

+++ NAV/Uninstall-NAVAll.ps1
@@ -52,10 +52,7 @@ function Uninstall-NAVAll {
     )
     begin {
         Import-SqlPsModule
-        if(!(Get-Module WebAdministration))
-        {
-            Import-Module WebAdministration -ErrorAction SilentlyContinue
-        }
+        
     }
     process {
         if(-Not(Test-Path $LogPath))
@@ -157,22 +154,38 @@ function Uninstall-BySetup ([string]$LogPath) {
 
 function Remove-AllNAVWebSites
 {
-    Write-Log "Remove all nav web sites"
-    $allNAVSites = Get-ChildItem "IIS:\Sites"  | Where-Object {$_.Name -match "NAV"}
-    foreach($navSite in $allNAVSites)
+    if(!(Get-Module WebAdministration))
     {
-        Write-Log "Remove web sites:  $navSite.name"
-        [System.IO.Directory]::Delete($navSite.PhysicalPath, $true)
-        Remove-Website -Name $navSite.name
+        Import-Module WebAdministration -ErrorAction SilentlyContinue
     }
 
+    Write-Log "Remove all nav web sites"
+    $allNAVSites = Get-ChildItem "IIS:\Sites"  | Where-Object {$_.Name -match "NAV"}
+    if($allNAVSites)
+    {
+        foreach($navSite in $allNAVSites)
+        {
+            Write-Log "Remove web sites:  $navSite.name"
+            if(Test-Path $navSite.PhysicalPath)
+            {
+                [System.IO.Directory]::Delete($navSite.PhysicalPath, $true)
+            }
+           
+            Remove-Website -Name $navSite.name
+        }
+    }
+    
     Write-Log "Remove all nave web application pools"
     $allNAVAppPools = Get-ChildItem "IIS:\AppPools" | Where-Object {$_.Name -match "NAV"} 
-    foreach($navAppPool in $allNAVAppPools)
+    if($allNAVAppPools)
     {
-        Write-Log "Remove nav web app pool: $navAppPool.name"
-        Remove-WebAppPool -Name $navAppPool.name 
+        foreach($navAppPool in $allNAVAppPools)
+        {
+            Write-Log "Remove nav web app pool: $navAppPool.name"
+            Remove-WebAppPool -Name $navAppPool.name 
+        }
     }
+    
 
     IISReset
 }
